[FIX] torch tensor in initialization + .csv

+++ leaspy/models/multivariate_model.py
@@ -15,7 +15,7 @@ class MultivariateModel(AbstractMultivariateModel):
         for k in parameters.keys():
             if k in ['mixing_matrix']:
                 continue
-            self.parameters[k] = torch.tensor(parameters[k])
+            self.parameters[k] = torch.tensor(parameters[k], dtype=torch.float32)
         self.attributes = AttributesFactory.attributes(self.name, self.dimension, self.source_dimension)
         self.attributes.update(['all'], self.parameters)
